Move getOnlinePlayers to a new Utils class
